not breaking the computational graph

+++ eg3d/training/face_id.py
@@ -1,21 +1,3 @@
-# https://github.com/timesler/facenet-pytorch/blob/master/models/inception_resnet_v1.py
-
-# MIT License
-
-# Copyright (c) 2019 Timothy Esler
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software 
-# and associated documentation files (the "Software"), to deal in the Software without restriction, 
-# including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, 
-# and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, 
-# subject to the following conditions:
-# The above copyright notice and this permission notice shall be included in all copies or substantial 
-# portions of the Software.
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT 
-# NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. 
-# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, 
-# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE 
-# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
 import os
 import torch
 from torch import nn
@@ -37,7 +19,7 @@ class FaceIDLoss:
         return feature_vector
 
     def transform_to_RGB(self, img):
-        img255 = (img * 127.5 + 128).clamp(0, 255).to(torch.uint8)
+        img255 = (img * 127.5 + 128).clamp(0, 255)
 
         return img255.float()
 
@@ -55,6 +37,24 @@ class FaceIDLoss:
     def get_loss(self, latent_chords, new_latent_chords):
         pass
 
+
+# MIT License
+
+# Copyright (c) 2019 Timothy Esler
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this software 
+# and associated documentation files (the "Software"), to deal in the Software without restriction, 
+# including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+# and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, 
+# subject to the following conditions:
+# The above copyright notice and this permission notice shall be included in all copies or substantial 
+# portions of the Software.
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT 
+# NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. 
+# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, 
+# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE 
+# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+# https://github.com/timesler/facenet-pytorch/blob/master/models/inception_resnet_v1.py
 class BasicConv2d(nn.Module):
     def __init__(self, in_planes, out_planes, kernel_size, stride, padding=0):
         super().__init__()
